Fix runtime crash
