+++ app/build.gradle.kts
@@ -135,5 +135,6 @@ dependencies {
     debugImplementation(libs.leak.canary)
 
     implementation(libs.hilt.android)
+    implementation(libs.androidx.hilt.navigation.compose)
     ksp(libs.hilt.compiler)
 }
