+++ app/src/main/java/com/august/jetcaster/ui/JetcasterApp.kt
@@ -22,6 +22,7 @@ import androidx.compose.material.TextButton
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -49,15 +50,8 @@ fun JetcasterApp(
                     }
                 )
             }
-            composable(Screen.Player.route) { backStackEntry ->
-                val playerViewModel: PlayerViewModel = viewModel(
-                    factory = PlayerViewModel.provideFactory(
-                        owner = backStackEntry,
-                        defaultArgs = backStackEntry.arguments
-                    )
-                )
+            composable(Screen.Player.route) {
                 PlayerScreen(
-                    playerViewModel,
                     windowSizeClass,
                     displayFeatures,
                     onBackPress = appState::navigateBack
