+++ app/src/main/java/com/august/jetcaster/ui/home/category/PodcastCategoryViewModel.kt
@@ -16,18 +16,12 @@
 
 package com.august.jetcaster.ui.home.category
 
-import android.content.Context
 import android.os.Bundle
 import androidx.lifecycle.AbstractSavedStateViewModelFactory
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.createSavedStateHandle
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
 import androidx.savedstate.SavedStateRegistryOwner
-import com.august.jetcaster.Graph
 import com.august.jetcaster.data.CategoryStore
 import com.august.jetcaster.data.EpisodeToPodcast
 import com.august.jetcaster.data.PodcastStore
@@ -41,9 +35,6 @@ import javax.inject.Inject
 
 @HiltViewModel
 class PodcastCategoryViewModel @Inject constructor(
-    // FIXME: private val categoryId: Long,
-    private val categoryStore: CategoryStore,
-    private val podcastStore: PodcastStore,
     savedStateHandle: SavedStateHandle,
 ) : ViewModel() {
     private val _state = MutableStateFlow(PodcastCategoryViewState())
@@ -51,8 +42,12 @@ class PodcastCategoryViewModel @Inject constructor(
     val state: StateFlow<PodcastCategoryViewState>
         get() = _state
 
+    @Inject lateinit var categoryStore: CategoryStore
+    @Inject lateinit var podcastStore: PodcastStore
+
+
     init {
-        val categoryId = savedStateHandle["category_id_key"] ?: 0L
+        val categoryId = savedStateHandle["categoryId"] ?: 0L
 
         viewModelScope.launch {
             val recentPodcastsFlow = categoryStore.podcastsInCategorySortedByPodcastCount(
@@ -82,8 +77,20 @@ class PodcastCategoryViewModel @Inject constructor(
     }
 
     companion object {
-        fun provideFactory(owner: Context): ViewModelProvider.Factory? {
-            TODO("Not yet implemented")
+        fun provideFactory(
+            owner: SavedStateRegistryOwner,
+            defaultArgs: Bundle? = null,
+        ): AbstractSavedStateViewModelFactory {
+            return object : AbstractSavedStateViewModelFactory(owner, defaultArgs) {
+                @Suppress("UNCHECKED_CAST")
+                override fun <T : ViewModel> create(
+                    key: String,
+                    modelClass: Class<T>,
+                    handle: SavedStateHandle
+                ): T {
+                    return PodcastCategoryViewModel(handle) as T
+                }
+            }
         }
     }
 }
